Diabetes Prediction using Random Forest
This is a project in which a dataset containing information about patients tested with diabetes, aiming to predict if a certain patient will have diabetes or not. For this purpose, after the data is preprocessed, a decision tree and a Random Forest are built using K-fold cross-validation with 5 folds, plotting also parameters such as accuracy and specificity of the model.